# internal/swagger/swagger.yaml
definitions:
  errors.Error:
    properties:
      code:
        type: integer
      detail:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
  schema.Menu:
    properties:
      children:
        description: Child menus
        items:
          $ref: '#/definitions/schema.Menu'
        type: array
      code:
        description: Code of menu (unique for each level)
        type: string
      created_at:
        description: Create time
        type: string
      description:
        description: Details about menu
        type: string
      id:
        description: Unique ID
        type: string
      name:
        description: Display name of menu
        type: string
      parent_id:
        description: Parent ID (From Menu.ID)
        type: string
      parent_path:
        description: Parent path (split by .)
        type: string
      path:
        description: Access path of menu
        type: string
      properties:
        description: Properties of menu (JSON)
        type: string
      resources:
        description: Resources of menu
        items:
          $ref: '#/definitions/schema.MenuMeta'
        type: array
      sequence:
        description: Sequence for sorting (Order by desc)
        type: integer
      status:
        description: Status of menu (enabled, disabled)
        type: string
      type:
        description: Type of menu (page, button)
        type: string
      updated_at:
        description: Update time
        type: string
    type: object
  schema.MenuForm:
    properties:
      code:
        description: Code of menu (unique for each level)
        maxLength: 32
        type: string
      description:
        description: Details about menu
        type: string
      name:
        description: Display name of menu
        maxLength: 128
        type: string
      parent_id:
        description: Parent ID (From Menu.ID)
        type: string
      path:
        description: Access path of menu
        type: string
      properties:
        description: Properties of menu (JSON)
        type: string
      resources:
        description: Resources of menu
        items:
          $ref: '#/definitions/schema.MenuMeta'
        type: array
      sequence:
        description: Sequence for sorting (Order by desc)
        type: integer
      status:
        description: Status of menu (enabled, disabled)
        enum:
        - disabled
        - enabled
        type: string
      type:
        description: Type of menu (page, button)
        enum:
        - page
        - button
        type: string
    required:
    - code
    - name
    - status
    - type
    type: object
  schema.MenuMeta:
    properties:
      created_at:
        description: Create time
        type: string
      id:
        description: Unique ID
        type: string
      menu_id:
        description: From Menu.ID
        type: string
      method:
        description: HTTP method
        type: string
      path:
        description: API request path (e.g. /api/v1/users/:id)
        type: string
      updated_at:
        description: Update time
        type: string
    type: object
...
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
info:
  contact: {}
  description: A lightweight, flexible, elegant and full-featured RBAC scaffolding
    based on GIN + GORM 2.0 + Casbin 2.0 + Wire DI.
  title: ginadmin
  version: v10.1.0
paths:
...
  /api/v1/menus:
    get:
      parameters:
      - description: Code path of menu (like xxx.xxx.xxx)
        in: query
        name: code
        type: string
      - description: Name of menu
        in: query
        name: name
        type: string
      - description: Whether to include menu resources
        in: query
        name: includeResources
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Menu'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query menu tree data
      tags:
      - MenuAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.MenuForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create menu record
      tags:
      - MenuAPI
  /api/v1/menus/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete menu record by ID
      tags:
      - MenuAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Menu'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get menu record by ID
      tags:
      - MenuAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.MenuForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update menu record by ID
      tags:
      - MenuAPI
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"